Keep file response disconnect tests in test_responses.py
